refactor jest config to use babel 7 & projects
