+++ babel.config.js
@@ -1,0 +1,12 @@
+
+module.exports = {
+    presets: [
+        [ '@babel/preset-env', {
+            targets: { node: 'current' }
+        } ],
+        [ '@babel/preset-typescript', {
+            allowNamespaces: true
+        } ],
+        '@babel/preset-react'
+    ],
+};
