+++ package.json
@@ -28,7 +28,10 @@
     "packages/front/node_modules"
   ],
   "devDependencies": {
-    "@babel/core": "^7.12.7",
+    "@babel/core": "^7.12.10",
+    "@babel/preset-env": "^7.12.11",
+    "@babel/preset-react": "^7.12.10",
+    "@babel/preset-typescript": "^7.12.7",
     "@material-ui/core": "^4.9.11",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.50",
@@ -37,12 +40,13 @@
     "@storybook/addon-links": "^6.1.2",
     "@storybook/react": "^6.1.2",
     "@testing-library/react-hooks": "^3.7.0",
-    "@types/jest": "^25.2.1",
+    "@types/jest": "^26.0.19",
     "@types/react": "^16.9.53",
+    "babel-jest": "^26.6.3",
     "babel-loader": "^8.2.1",
     "copyfiles": "^2.4.1",
     "eslint": "^7.4.0",
-    "jest": "^25.5.2",
+    "jest": "^26.6.3",
     "pixi.js": "^5.2.1",
     "react": "^17.0.1",
     "react-dom": "17.0.1",
