+++ packages/animated-complex-sprite/package.json
@@ -12,5 +12,8 @@
   },
   "dependencies": {
     "@pixi/filter-outline": "^3.1.1"
+  },
+  "devDependencies": {
+    "@timeflies/devtools": "1.0.0"
   }
 }
