+++ packages/back/jest.config.js
@@ -1,11 +1,11 @@
 // For a detailed explanation regarding each configuration property, visit:
 // https://jestjs.io/docs/en/configuration.html
 
-const rootConfig = require('../../jest.config');
+const rootConfig = require('../../jest.config.base');
 
 module.exports = {
 
-  ...rootConfig,
+  ...rootConfig(__dirname),
 
   setupFiles: [
     '<rootDir>/src/setupTestsEnv.ts'
